Put integration test (without frontend) in separate package vs. selenium test
