--- cucumbertests/src/test/java/eu/luminis/breed/sleuthzipkin/RunTest.java
@@ -1,11 +1,0 @@
-package eu.luminis.breed.sleuthzipkin;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.junit.Cucumber;
-import org.junit.runner.RunWith;
-
-@RunWith(Cucumber.class)
-@CucumberOptions(format = {"pretty", "html:target/cucumber"}, features="src/test/resources/")
-public class RunTest {
-
-}
