+++ cucumbertests/src/test/java/eu/luminis/breed/sleuthzipkin/integration/BreakFastServiceApi.java
@@ -1,4 +1,4 @@
-package eu.luminis.breed.sleuthzipkin;
+package eu.luminis.breed.sleuthzipkin.integration;
 
 import static org.junit.Assert.fail;
