+++ cucumbertests/src/test/java/eu/luminis/breed/sleuthzipkin/integration/RunTest.java
@@ -1,0 +1,11 @@
+package eu.luminis.breed.sleuthzipkin.integration;
+
+import cucumber.api.CucumberOptions;
+import cucumber.api.junit.Cucumber;
+import org.junit.runner.RunWith;
+
+@RunWith(Cucumber.class)
+@CucumberOptions(glue = "eu.luminis.breed.sleuthzipkin.integration",format = {"pretty", "html:target/cucumber"}, features="src/test/resources/")
+public class RunTest {
+
+}
